Explicit cast from android::Base::Errno to StatusCode
Previously, Result<T>.error().code() returned int and this was
convertible to StatusCode which is an enum class based on int32_t.

However, this actually was type unsafe because the return type of
error().code() was actually mean to be that of errno. errno and
StatusCode are both int, but their semantis are different.

To handle such problems, Result<T>.error().code() now returns
android::base::Errno which is a wrapper to errno, but not to other
int-based types like status_t and StatusCode.

As a (happy) side effect, a direct conversion from Errno to StatusCode
has became impossible. To work around this, error().code() is explicitly
casted to int so that it can then be converted into StatusCode.

Eventually, Result<T> used here should be changed into Result<T,
StatusCodeWrapper> where StatusCodeWrapper is a wrapper class to
StatusCode that implements the contract to be used as the error type in
android::base::Result.

Bug: 209929099
Test: m

+++ automotive/vehicle/2.0/default/impl/vhal_v2_0/EmulatedVehicleHal.cpp
@@ -157,7 +157,7 @@ VehicleHal::VehiclePropValuePtr EmulatedVehicleHal::get(
                 const auto& ret = mEmulatedUserHal->onGetProperty(requestedPropValue);
                 if (!ret.ok()) {
                     ALOGE("get(): User HAL returned error: %s", ret.error().message().c_str());
-                    *outStatus = StatusCode(ret.error().code());
+                    *outStatus = StatusCode(static_cast<int>(ret.error().code()));
                 } else {
                     auto value = ret.value().get();
                     if (value != nullptr) {
